+++ Changes.md
@@ -1,3 +1,5 @@
+* Allow one to save a modified schedule from the SchedulerUI
+* #1096 - Allow one to optimize tables in groups
 * #1102 - Send to web sockets in the background to speed up web clients
 * automatically select common columns when loading a schedule into the scheduler UI, website and when uploading teams
 * #1100 - Switch the scoreboard to web sockets to avoid issues with web servers